Row and Columns now use a lambda callback for iteration
Tried using an Iterator pattern but the underlyding cpp driver
code does not allow for copying iterators -- thus a user could
easily invalidate their iterator without much effort and cause
problems.  By forcing at least the Rows in a Result to be
iterated over one at a time in a lambda function with a const& Row
which the user cannot move/copy it forces the iteration to happen
in the correct order.  Columns can still be grabbed via index
or name but a ForEachColumn(lambda) was added for consistency.